pkg/alertmanager: define all possible receivers
If the user-defined Alertmanager configuration declared a receiver that
isn't yet supported by AlertmanagerConfig (e.g. email, Pushover or
VictorOps), the receiver declaration would be dropped from the generated
configuration.

Closes #3665
